Agent AI/requirements.txt
# Agent AI Requirements
# Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0

# OpenAI API
openai==1.3.0

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0

# Environment
python-dotenv==1.0.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Logging
colorlog==6.7.0

# Development
pytest==7.4.3
pytest-flask==1.2.0
